+++ udata/core/dataset/csv.py
@@ -37,7 +37,7 @@ class DatasetCsvAdapter(csv.Adapter):
         ("archived", lambda o: o.archived or False),
         ("resources_count", lambda o: len(o.resources)),
         ("main_resources_count", lambda o: len([r for r in o.resources if r.type == "main"])),
-        ("resources_formats", lambda o: ",".join(set(r.format for r in o.resources))),
+        ("resources_formats", lambda o: ",".join(set(r.format for r in o.resources if r.format))),
         "downloads",
         ("harvest.backend", lambda r: r.harvest and r.harvest.backend),
         ("harvest.domain", lambda r: r.harvest and r.harvest.domain),