+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current (in progress)
 
+- Add migration to delete duplicate resources due to ODS harvesting [#3158](https://github.com/opendatateam/udata/pull/3158)
 - Add discussion.posted_on in discussion sort choices [3168](https://github.com/opendatateam/udata/pull/3168)
 
 ## 9.2.2 (2024-10-08)